Relax gem versions
We started out with this meta-gem strictly setting the versions
down to the tiny version. In practice, this means there are lots
of updates needed to keep the gem viable. This assumes that these
gem authors are following semver conventions. Easier to come back
and lock down a specific gem if we find an issue.