Remove ci configuration and template generation
Rationale: thincloud-test is used *mostly* as the underpinning for
thincloud-test-rails. Rails projects, for New Leaders, don't often
have a CI component. That means we have this configuration floating
around in projects that don't need it.

This removal is in preparation to extract the ci templates into a
thincloud-ci gem which can be added as needed.