Remove metrics configuration and related files
Rationale: We currently include Simplecov and Cane as dependencies
of thincloud-test. Rather than have a few metrics tools buried in
the test gem. When these aren't needed they end up leaving
configuration and dependency liabilities in the target project.

This removal is in preparation to extract our metric tools into
a thincloud-metrics gem that can be added as desired.